Add two different persistence clients to save/restore state
FlatFile will just save to a state file on the local file system
designated by the persistence_state_file config option, and
the Redis one will write to a key in redist designated by the
persistence_state_key config. The redis host, port, and password as well
as Redix child spec are configured in config/dev.secrets.exs (I put a
template for this in config/dev.secrets.exs.example, and gitignored the
real file) or in config/releases.exs for the production release, which
will read the values from ENV variables.

For now I am using Redis for dev and production, and the flat file for
the test environment. It would make sense to use flat file for dev too,
but I wanted to be able to verify the redis persistence client worked
properly, so I switched it over for now.