Store updated_at and saved_at in state, and add persistence functions
Added save_state/1, load_state/1, and get_updated_at/1 to Data. So we
can persist the current state and re-load it on app startup as needed,
and we can read out when the data was last updated so we can determine
when we should check next, based on the configured intervals.