Add DataManager to handle saving/loading/updating Data state
This basically starts on application start and loads the data from state
if it's available, and figures out based on the updated_at value in the
state when to schedule the next update for. It attempts to update about
once an hour, even if the app crashes and gets re-started. Each time it
successfully updates the data from the PVA website, it saves the state to
the persistence store, so the state in memory and on disk should
theoretically stay in fairly close sync.