Replace all the division data instead of putting each division
The previous implementation had a bug where if there were divisions in
the previous season that didn't exist in the new season, they would
still stick around in the divisions map because they weren't being
overwritten. This just replaces the whole divisions map in one go, which
is kind of a better API for the DataManager to use anyway.