Fix rank numbers when ties are involved
If more than one team occupies a given rank, the next teams below them
should be ranked at the appropriate level with the assumption that the
tie will be broken and all the intermediate ranks will be occupied. For
example, if two teams are tied for first place, the next team in the
standings should be in third, not second.