Compare functions return numeric values, compare_teams returns bool
I realized that for the purposes of the implementation of the
compare_teams function, I needed the individual comparision functions to
be able to distinguish between a tie and greater than, unlike the
boolean versions of those functions. That way I can continue comparing
if we have a tie, but pass through if we have a greater than or less
than result. The compare_teams function then just translates the final
numeric result into the boolean result that Enum.sort expects.