Add test for Division.compare_teams/2 using fixture data
This is not a perfect test, as the data is real data that does not test
all the edge cases, but it does sort it in the correct order. It would
be ideal to construct a test that tests all the edge cases but hopefully
the individual tests combined with this overall test will cover it
reasonably well.