Upgrade some versions of things including erlang/elixir
The main motivation here was because I couldn't get the
meeseeks_html5ever NIF compiled on M1, and so I wanted to update that to
a newer version that did support M1. That deprecated some older erlang
and elixir versions, so I upgraded to the versions that did explicitly
support. I can't recall why I updated the other dependencies, but so far
they seem to be working, so nice to have them updated. I still need to
do some more thorough testing of the full app, though.