Update to support the Team Sideline version of the website
Team Sideline unsurprisingly doesn't have the concept of "match points"
that PVA perhaps invented or used for many years. Instead they use
"average point differential" as their secondary ranking mechanism behind
winning percentage. So, I've removed all references to match points and
added in average point differential to the standings info.

The site is also organized differently, where the landing page shows a
list of divisions and links to those divisions, and then each of those
pages has all of the data for that division (standings, scores, and
schedules). So, the new scraper first hits the page that lists the
divisions, grabs that list, and then hits each division, parsing that
entirely and computing standings and ranks in the process. We only rely
on the standings on the website for the list of teams, but all the data
is calculated from the scores, which feels right.

Unfortunately as it currently stands, the sort order in the standings on
the portlandvolleyball.org does not match the order that this code
outputs, but this code attempts to follow the tie breaking rules as
stated on the website, and I hope eventually the website will be updated
to match the stated rules, or in the case that I am misinterpreting the
rules, I will learn how and update my side.

I marked a few tests as skipped which are related to the team comparison
functions and adding ranks. I don't want to spend too much time covering
those with tests independently until the sort order and rule list
stabilizes. I replaced the existing fixtures with a snapshot of the
relevant pages at the time when I started working on this. Over time
maybe I'll also add snapshots of the beginning of the season and/or when
there are unique tie-breaker scenearios that need to be tested.