Check out Women's Monday B for some weird stuff on 2025-03-10
There was a match where they only played two sets and the ref
(presumably) left a comment to that effect. Then there was a match where
one team forfeited. I'm not sure if that is a standard system message
that is reliable or if it's just a comment, but the parser was not able
to parse the forfeit from what I can see. It just ignored it. It was
able to parse the partial match just fine apparently, though I'm not
sure if it catured the ref since the ref was in the same span. That
doesn't show on the alt-pva website currently for scores, but I think
the expectation is that it should be parsed. I need to write tests for
these cases and makes sure I'm handling it correctly.