Make team parsing more flexible
portlandvolleyball.org added more columns between the team name and the
coach column, which broke the parsing. This allows for a flexible number
of columns between them as long as the name is in the second column and
the coach is in the last column. This will no longer fail if the format
changes but we might get incorrect data. Might be worth adding a rollbar
warning if the number of columns is unexpected, but for now I'll just
get it working again.